drivers: ieee802154_nrf5: Use BLE.ADDR instead of FICR to create EUI64
Use BLE.ADDR to create unique (to some extent) EUI64
in some cases inside the IEEE 802.15.4 driver.

The amount of EUI64-s available in such a way is very limited (~16 million).
However, currently there does not seem to be another feasible way to
get a device identifiers (such are kept in SICR, to which the
radio core has no access).

Signed-off-by: Piotr Koziar <piotr.koziar@nordicsemi.no>